Fixed Memory representation issue.

+++ src/Components/DataMemory.java
@@ -3,7 +3,7 @@ package Components;
 //Only used in Single Cycle / Pipe architectures
 public class DataMemory extends MemoryModule{
 
-    public DataMemory(int wordsize, int size, int offset){
+    public DataMemory(Double wordsize, int size, long offset){
         super(wordsize, size, offset);
     }
 }
