+++ src/Components/InstructionMemory.java
@@ -3,7 +3,7 @@ package Components;
 public class InstructionMemory extends MemoryModule{
     private ProgramCounter pc;
 
-    public InstructionMemory(int wordsize, int size, int offset){
+    public InstructionMemory(Double wordsize, int size, long offset){
         super(wordsize, size, offset);
         this.pc = new ProgramCounter(wordsize);
     }
