+++ src/Components/RegisterFile.java
@@ -1,7 +1,7 @@
 package Components;
 
 public class RegisterFile extends MemoryModule{
-    public RegisterFile(int wordsize, int registeramount){
+    public RegisterFile(Double wordsize, int registeramount){
         super(wordsize, registeramount, 0); //Registers have no offset.
     }
 }
